Finished implementing PegOut. To test
