--- service-kit/src/main/scala/co/topl/brambl/playground/BridgeEntity.scala
@@ -1,5 +1,0 @@
-package co.topl.brambl.playground
-
-trait BridgeEntity extends BitcoinWallet with ToplWallet {
-  override val walletName: String
-}
